chore: reorg relay code
